requirements.txt
# This file is here because many Platforms as a Service look for
# requirements.txt in the root directory of a project.

django-import-export==0.5.1
-r requirements/production.txt
